requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# Discord Bot
discord.py>=2.3.2
# Alternative: py-cord>=2.4.1

# Web Scraping
playwright>=1.40.0
playwright-stealth>=1.0.6

# Utilities
aiohttp>=3.9.0
python-dateutil>=2.8.2
